backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
python-multipart
pydantic
python-dotenv